requirements.txt
albumentations==1.3.1
matplotlib>=3.8.4
numpy>=1.26.4
opencv-python>=4.10.0.84
pandas>=2.2.2
pycocotools>=2.0.8
PyYAML>=6.0.2
pytest>=8.3.2
tqdm>=4.66.5
ultralytics>=8.3.0
